Introduce par_iter for faster proof gen
Use parallel iteration to shave off a large part of the proof generation
time. The most expensive part is the range proof on the bits of the
liability proof since it has to generate so many which is why this helps
so much.

**Before**

gen-solvency/10x10      time:   [2.1260 s 2.1565 s 2.2788 s]

**After**

gen-solvency/10x10      time:   [340.99 ms 341.31 ms 342.60 ms]
                        change: [-85.036% -84.480% -83.885%] (p = 0.09 > 0.05)